add more io functions

+++ cmd/lib/kernel_pthread_attr.go
@@ -15,7 +15,7 @@ import (
 func libKernel_scePthreadAttrInit(attrHandlePtr uintptr) uintptr {
 	err := libKernel_pthread_attr_init(attrHandlePtr)
 	if err != 0 {
-		return uintptr(uint32(err) - 0x7FFE0000)
+		return err - 0x7FFE0000
 	}
 
 	return 0
